+++ section-2/step-04/src/main/java/com/carmanagement/agentic/workflow/FeedbackWorkflow.java
@@ -22,7 +22,7 @@ public interface FeedbackWorkflow {
             String carMake,
             String carModel,
             Integer carYear,
-            Long carNumber,
+            Integer carNumber,
             String carCondition,
             String rentalFeedback,
             String cleaningFeedback,
